refactor(content-api): Update RawQueryBuilder and CollectionBuilder for improved query handling
