--- core-web/libs/sdk/client/src/lib/client/content/builders/base/index.ts
@@ -1,1 +1,0 @@
-export { BaseBuilder } from './base-builder';
