+++ core-web/libs/sdk/client/src/lib/client/content/builders/rawQuery/raw-query-builder.ts
@@ -6,8 +6,7 @@ import {
     DotErrorContent
 } from '@dotcms/types';
 
-import { shouldAddSiteIdConstraint } from '../../shared/utils';
-import { BaseBuilder } from '../base';
+import { BaseBuilder } from '../base/base-builder';
 import { sanitizeQuery } from '../query/utils';
 
 /**
@@ -28,24 +27,6 @@ export class RawQueryBuilder<T = unknown> extends BaseBuilder<T> {
     #rawQuery: string;
 
     /**
-     * Language ID for filtering content.
-     * @private
-     */
-    #languageId: number | string = 1;
-
-    /**
-     * Whether to retrieve draft content.
-     * @private
-     */
-    #draft = false;
-
-    /**
-     * Variant ID for experiments.
-     * @private
-     */
-    #variantId?: string;
-
-    /**
      * Creates an instance of RawQueryBuilder.
      *
      * @param {object} params - Constructor parameters
@@ -63,70 +44,6 @@ export class RawQueryBuilder<T = unknown> extends BaseBuilder<T> {
     }) {
         super(params);
         this.#rawQuery = params.rawQuery;
-    }
-
-    /**
-     * Filters the content by the specified language ID.
-     *
-     * @example
-     * ```typescript
-     * const response = await client.content
-     *     .query('+contentType:Blog')
-     *     .language(1)
-     *     .limit(10);
-     * ```
-     *
-     * @param {number | string} languageId - The language ID to filter the content by
-     * @return {this} The builder instance for method chaining
-     * @memberof RawQueryBuilder
-     */
-    language(languageId: number | string): this {
-        this.#languageId = languageId;
-
-        return this;
-    }
-
-    /**
-     * Retrieves draft content instead of live content.
-     *
-     * @example
-     * ```typescript
-     * const response = await client.content
-     *     .query('+contentType:Blog')
-     *     .draft()
-     *     .limit(10);
-     * ```
-     *
-     * @return {this} The builder instance for method chaining
-     * @memberof RawQueryBuilder
-     */
-    draft(): this {
-        this.#draft = true;
-
-        return this;
-    }
-
-    /**
-     * Filters the content by a variant ID for Experiments.
-     *
-     * More information here: {@link https://www.dotcms.com/docs/latest/experiments-and-a-b-testing}
-     *
-     * @example
-     * ```typescript
-     * const response = await client.content
-     *     .query('+contentType:Blog')
-     *     .variant('legends-forceSensitive')
-     *     .limit(10);
-     * ```
-     *
-     * @param {string} variantId - A string that represents a variant ID
-     * @return {this} The builder instance for method chaining
-     * @memberof RawQueryBuilder
-     */
-    variant(variantId: string): this {
-        this.#variantId = variantId;
-
-        return this;
     }
 
     /**
@@ -159,65 +76,28 @@ export class RawQueryBuilder<T = unknown> extends BaseBuilder<T> {
     }
 
     /**
-     * Gets the language ID for the query.
+     * Raw queries do not automatically inject any system constraints.
+     * If you need constraints like language, live/draft, site, or variant, include them in the raw query string.
      *
      * @protected
-     * @return {number | string} The language ID
+     * @return {undefined} No languageId is sent unless provided in the raw query string itself.
      * @memberof RawQueryBuilder
      */
-    protected getLanguageId(): number | string {
-        return this.#languageId;
+    protected getLanguageId(): undefined {
+        return undefined;
     }
 
     /**
-     * Builds the final Lucene query string by adding essential system constraints.
+     * Builds the final Lucene query string.
      *
-     * This method takes the raw query and adds:
-     * 1. Language ID constraint
-     * 2. Live/draft status constraint
-     * 3. Variant constraint (if specified)
-     * 4. Site ID constraint (if conditions are met)
-     *
-     * The raw query is used as-is with minimal sanitization (only trimming spaces).
-     * NO field prefixing is applied (unlike CollectionBuilder).
+     * Raw query is used as provided (only minimal sanitization is applied).
+     * No implicit constraints are added.
      *
      * @protected
      * @return {string} The final Lucene query string
      * @memberof RawQueryBuilder
-     *
-     * @example
-     * // Input: '+contentType:Blog +title:*'
-     * // Output: '+contentType:Blog +title:* +languageId:1 +live:true +conhost:123'
-     *
-     * @example
-     * // With draft flag:
-     * // Input: '+contentType:Blog'
-     * // Output: '+contentType:Blog +languageId:1 +(live:false AND working:true AND deleted:false) +conhost:123'
      */
     protected buildFinalQuery(): string {
-        // Start with the raw query (minimal sanitization only)
-        let query = sanitizeQuery(this.#rawQuery);
-
-        // Add language constraint
-        query = `${query} +languageId:${this.#languageId}`;
-
-        // Add draft/live constraint
-        if (this.#draft) {
-            query = `${query} +(live:false AND working:true AND deleted:false)`;
-        } else {
-            query = `${query} +live:true`;
-        }
-
-        // Add variant constraint if specified
-        if (this.#variantId) {
-            query = `${query} +variant:${this.#variantId}`;
-        }
-
-        // Add site constraint if needed (following same logic as CollectionBuilder)
-        if (shouldAddSiteIdConstraint(query, this.siteId)) {
-            query = `${query} +conhost:${this.siteId}`;
-        }
-
-        return sanitizeQuery(query);
+        return sanitizeQuery(this.#rawQuery);
     }
 }
